staticwaves_pod/requirements.txt
flask
requests
python-dotenv
rembg
pillow
onnxruntime
tenacity
pandas
openpyxl
